feat: add thunk and action typing
